rafactor: Modularize code.py and Indentation fix
