+++ code/helper.py
@@ -1,0 +1,84 @@
+import re
+import json
+import os
+
+spend_categories = ['Food', 'Groceries', 'Utilities', 'Transport', 'Shopping', 'Miscellaneous']
+
+spend_display_option = ['Day', 'Month']
+
+# set of implemented commands and their description
+commands = {
+    'menu': 'Display this menu',
+    'add': 'Record/Add a new spending',
+    'display': 'Show sum of expenditure for the current day/month',
+    'history': 'Display spending history',
+    'delete': 'Clear/Erase all your records',
+    'edit': 'Edit/Change spending details'
+}
+
+dateFormat = '%d-%b-%Y'
+timeFormat = '%H:%M'
+monthFormat = '%b-%Y'
+
+
+# function to load .json expense record data
+def read_json():
+    global user_list
+    try:
+        if os.stat('expense_record.json').st_size != 0:
+            with open('expense_record.json') as expense_record:
+                expense_record_data = json.load(expense_record)
+            user_list = expense_record_data
+
+    except FileNotFoundError:
+        print("---------NO RECORDS FOUND---------")
+
+
+def write_json(user_list):
+    try:
+        with open('expense_record.json', 'a') as json_file:
+            json.dump(user_list, json_file, ensure_ascii=False, indent=4)
+    except FileNotFoundError:
+        print('Sorry, the data file could not be found.')
+
+
+def validate_entered_amount(amount_entered):
+    if len(amount_entered) > 0 and len(amount_entered) <= 15:
+        if amount_entered.isdigit:
+            if re.match("^[0-9]*\\.?[0-9]*$", amount_entered):
+                amount = round(float(amount_entered), 2)
+                if amount > 0:
+                    return str(amount)
+    return 0
+
+
+# TODO: load from expense_record.json
+def getUserHistory(chat_id):
+    global user_list
+    if (str(chat_id) in user_list):
+        return user_list[str(chat_id)]
+    return None
+
+
+def getSpendCategories():
+    return spend_categories
+
+
+def getSpendDisplayOptions():
+    return spend_display_option
+
+
+def getCommands():
+    return commands
+
+
+def getDateFormat():
+    return dateFormat
+
+
+def getTimeFormat():
+    return timeFormat
+
+
+def getMonthFormat():
+    return monthFormat
