+++ code/history.py
@@ -1,0 +1,21 @@
+import helper
+
+
+# TODO: check if message is printed properly
+def run(message, bot):
+    try:
+        helper.read_json()
+        chat_id = message.chat.id
+        user_history = helper.getUserHistory(chat_id)
+        spend_total_str = ""
+        if user_history is None:
+            raise Exception("Sorry! No spending records found!")
+        spend_total_str = "Here is your spending history : \nDATE, CATEGORY, AMOUNT\n----------------------\n"
+        if len(user_history) == 0:
+            spend_total_str = "Sorry! No spending records found!"
+        else:
+            for rec in user_history:
+                spend_total_str += str(rec) + "\n"
+            bot.send_message(chat_id, spend_total_str)
+    except Exception as e:
+        bot.reply_to(message, "Oops!" + str(e))
