+++ test/test_start_and_menu_command.py
@@ -6,10 +6,11 @@ Created on Wed Sep 29 17:07:37 2021
 @author: deekay
 """
 
-import pytest
-#from code.code import start_and_menu_command
+# import pytest
+# from code.code import start_and_menu_command
+
 
 def test_start_and_menu_command_func():
-    #test_result = start_and_menu_command("/start")
-    #assert True == test_result, "Normal Case"
+    # test_result = start_and_menu_command("/start")
+    # assert True == test_result, "Normal Case"
     print("Hello")
